% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateAerodynamicResistance.R
\name{CalculateAerodynamicResistance}
\alias{CalculateAerodynamicResistance}
\title{CalculateAerodynamicResistance}
\usage{
CalculateAerodynamicResistance(
  FrictionVelocity_ms,
  ReferenceHeight_m,
  ZeroPlaneDisplacementHeight_m,
  RoughnessLength_m,
  MoninObukhovLength_m
)
}
\arguments{
\item{FrictionVelocity_ms}{Friction velocity in in m/s.}

\item{ZeroPlaneDisplacementHeight_m}{Displacement height in m.}

\item{RoughnessLength_m}{Roughness length in m.}

\item{MoninObukhovLength_m}{Monin-Obkukhiv length in m. Monin-Obkukhiv length for neutral stratification (Pasquill class D) is "infinity". This case is encoded by a value defined in GetConstants()$InfLength in this package.}
}
\value{
Aerodynamic resistance in s/m.
}
\description{
Calculates aerodynamic resistance according to Erisman and Draaijers (1995) page 58. equation 3.4.
}
\examples{

# Aerodynamic resistance for extremely unstable stratification
# over grassland
PasquillClass <- "A"
RoughnessLength_m <- 0.03
WindSpeed_ms <- 1.5
AnemometerHeight_m <- 10
ReferenceHeight_m <- 10
ZeroPlaneDisplacementHeight_m <- 7 * RoughnessLength_m

MOL_m <- CalculateMoninObukhovLength(
  PasquillClass = PasquillClass,
  RoughnessLength_m = RoughnessLength_m
)
FrictionVelocity_ms <- CalculateFrictionVelocity(
  WindSpeed_ms = WindSpeed_ms,
  AnemometerHeight_m = AnemometerHeight_m,
  ZeroPlaneDisplacementHeight_m = ZeroPlaneDisplacementHeight_m,
  RoughnessLength_m = RoughnessLength_m,
  MoninObukhovLength_m = MOL_m
)
CalculateAerodynamicResistance(
  FrictionVelocity_ms = FrictionVelocity_ms,
  ReferenceHeight_m = ReferenceHeight_m,
  ZeroPlaneDisplacementHeight_m = ZeroPlaneDisplacementHeight_m,
  RoughnessLength_m = RoughnessLength_m,
  MoninObukhovLength_m = MOL_m
)

}
\references{
Erisman JW, Draaijers GPJ. Atmospheric Deposition In Relation to Acidification and Eutrophication. 1995.
}
