% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateAirDensity.R
\name{CalculateAirDensity}
\alias{CalculateAirDensity}
\title{CalculateAirDensity}
\usage{
CalculateAirDensity(AirPressure_Pa, T_air_K)
}
\arguments{
\item{AirPressure_Pa}{Air pressure in Pa.}

\item{T_air_K}{Air temperature in K.}
}
\value{
Density of air in kg/m3.
}
\description{
Calculates the density of air according to Seinfeld and Pandis
(2006) page 735 eq. 16.36
}
\references{
Seinfeld JH, Pandis SN. Atmospheric Chemistry and Physics: From Air Pollution to Climate Change. Wiley; 2006.
}
