% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateDepositionVelocity.R
\name{CalculateDepositionVelocity}
\alias{CalculateDepositionVelocity}
\title{CalculateDepositionVelocity}
\usage{
CalculateDepositionVelocity(InputTable)
}
\arguments{
\item{InputTable}{A data frame with the following columns:

  - SunAngle_degree: The sun angle in degree. Only used to determine if its
  "day" (> 0°) or "night" (< 0°) which is required for the determination of
  the Pasquill stability class (GetPasquillClass()). The sun angle can easily
  be calculated with oce::sunAngle().

  - T_air_K: Air temperature in Kelvin.

  - AirPressure_Pa: Air pressure in Pa. Required for the calculation of the
  mean free path of an air molecule, which is required for the calculation of
  the particle settling (sedimentation) velocity.

  - GlobalRadiation_W_m2: Global radiation in W/m2. Required for the
  determination of the Pasquill stability class (GetPasquillClass()).

  - CloudCover_percent: Cloud cover in percent. Required for the
  determination of the Pasquill stability class (GetPasquillClass()).

  - WindSpeedAtAnemometerHeight_ms: Wind speed (m/s) at the the anemometer
  height. Required for GetPasquillClass() and CalculateFrictionVelocity().

  - RoughnessLength_m: Roughness length (m) of the land cover for which the
  anemometer wind speed is provided. Required for multiple functions.

  - ZeroPlaneDisplacementHeight_m: Zero plane displacement height (m) of the
  land cover for which the anemometer wind speed is provided. Required for
  multiple functions.

  - AnemometerHeight_m: Height to which the wind speed data refers

  - SurfaceIsWet_bool: Boolean indicating whether the surface is wet, in
  which case the particle rebound effect is disabled. Can be set e.g.
  depending on relative humidity and/or precipitation events.

  - SurfaceIsVegetated_bool: Boolean indicating whether the surface is
  vegetated. Should be TRUE for all LUCs currently implemented in ddpart.

  - LUCNames: Land use classes (character). See ?GetLandUseParameters for
  information.

  - ReferenceHeight_m: Aeordynamic resistance is calculated between
  ReferenceHeight_m and the effective receptor height (RoughnessLength_m +
  ZeroPlaneDisplacementHeight_m).

  - RoughnessLength_m: Roughness length of the land use class.

  - ZeroPlaneDisplacementHeight_m: Zero-plane displacement height of the land
   use class.

  - Season: Integer determining the season according to Zhang
  et al. (2001) table 2.

  - DryParticleDiameter_m: Particle diameter in m before the growth particles
  due to water uptake is taken into account (see parameter "AerosolType").

  - ParticleDensity_kgm3: Density of the particles in kg/m3.

  - AerosolType: ´Set parameter AerosolType to "Dry" to disable calculation of
  hygroscopic swelling. Set parameter "AerosolType" to one of "SeaSalt",
  "Urban", "Rural", "AmmoniumSulfate" to enable hygroscopic swelling
  according to Zhang et al. (2001). See CalculateHygroscopicSwelling() for
  further details.

  - RelHum_percent: Relativ humidity (%) for hygroscopic swelling of
  particles.

  - Parametrization: A character indicates which parametrization should be
  used. See ?GetLandUseParameters for allowed values.}
}
\value{
A data frame repeating the InputTable plus additional columns with
  calculated values.
}
\description{
Calculates the deposition velocity of particle according to the
  Zhang et al. (2001) and the Emerson et al. (2020) publications.

  This is a wrapper function over the various dry deposition sub-processes.
  It is designed for use cases where information on wind speed is
  available for the same land use type where deposition velocities should be
  calculated (e.g. wind speed measured at 10 m anemometer height and
  concentrations at 1.5 m, both over grassland.) Use
  CalculateDepositionVelocity2() for situations where wind speed data is
  available for one land use but dry deposition velocities should be
  calculated for another land use.

  Required inputs are:

  (1) basic meteorological data (wind speed, global ration, relative
  humidity, cloud cover, air temperature, air pressure, surface wetness)

  (2) information on the surface properties (land use class, roughness length
  and displacement height) and reference height (height of concentration
  measurements)

  (3) particle properties (dry particle diameter, density, optional: aerosol
  type to calculate hygroscopic swelling)

  The calculation steps in CalculateDepositionVelocity() are:

  1. Calculate meteorological basics (e.g. viscosity and density of air) and
  whether its day of night (based on parameter SunAngle_degree).

  2. Calculate the Pasquill class (general classification of atmospheric
  stability).

  3. Calculate other meteorological parameters (Monin-Obukhov length,
  friction velocity, stability corrections).

  4. Calculate the aerodynamic resistance (Ra) between the reference height
  and the effective height of the receptor (displacement height plus
  roughness length). Calculate surface resistance (Rs) and
  gravitation settling.

  5. Calculate dry deposition velocity.
}
\examples{
# See vignette.

}
\references{
Zhang L, Gong S, Padro J, Barrie L. A size-segregated particle
dry deposition scheme for an atmospheric aerosol module. Atmospheric
Environment 2001;35:549–560. Emerson EW, Hodshire AL, DeBolt HM, Bilsback KR,
Pierce JR, McMeeking GR, Farmer DK. Revisiting particle dry deposition and
its role in radiative effect estimates. Proceedings of the National Academy
of Sciences 2020;117:26076–26082.
}
