% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateDepositionVelocity2.R
\name{CalculateDepositionVelocity2}
\alias{CalculateDepositionVelocity2}
\title{CalculateDepositionVelocity2}
\usage{
CalculateDepositionVelocity2(InputTable)
}
\arguments{
\item{InputTable}{A data frame with the all columns required by
CalculateDepositionVelocity(), but:

  - RoughnessLengthAnemometer_m: Roughness length (m) of the land cover for
  which the anemometer wind speed is provided. Required for multiple
  functions.

  - ZeroPlaneDisplacementHeightAnemometer_m: Zero plane displacement height
  (m) of the land cover for which the anemometer wind speed is provided.
  Required for multiple functions.

  - WindSpeedBlendingHeight_m: Height in m to which the wind speed is
  extrapolated. At this height, the wind speed is assumed to be independent
  of the underlying land use.

  - TargetLUCCodeZhang2001: Integer determining the land use class of the
  target land use according to Zhang et al. (2001) table 2.

  - RoughnessLengthTargetLUC_m: Roughness length (m) of the land cover for
  which the dry deposition velocities are calculated.

  - ZeroPlaneDisplacementHeightTargetLUC_m: Zero plane displacement height
  (m) of the land cover for which the dry deposition velocities are
  calculated.}
}
\value{
A data frame repeating the InputTable plus additional columns with
  calculated values.
}
\description{
Calculates the deposition velocity of particle according to the
  Zhang et al. (2001) and the Emerson et al. (2020) publications.

  This is a wrapper function over the various dry deposition sub-processes.
  It is designed for use cases where information on wind speed is
  available for one land use but dry deposition velocities should be
  calculated for another land use. For example, wind speed is available from
  modelled data for standard WMO conditions (10 m over grassland) but the
  dry deposition velocity should be calculated for forest. Use
  CalculateDepositionVelocity() for the more typical case where wind speed
  information is available for the sane land use type where dry deposition
  should be calculated.

  The approach therefore differentiates between two sites:

  - "anemometer site": Location/land use class where the wind speed
  measurements are available. Characterized by parameters
  "RoughnessLengthAnemometer_m" and "ZeroPlaneDisplacementHeightAnemometer_m"

  - "target land use site": Location/land use class where the dry deposition
  velocity should be calculated. Characterized by parameters
  "RoughnessLengthTargetLUC_m", "ZeroPlaneDisplacementHeightTargetLUC_m",
  "Season"

  CalculateDepositionVelocity2() vertically extrapolates the wind speed at
  "anemometer site" the to a "blending height" (e.g. 50 m) where atmospheric
  conditions are somewhat independent of the underlying surface. From the
  blending height, the parameters relevant for particle dry deposition to the
  target land use site (friction velocity, aerodynamic resistance, etc.) are
  be calculated. The main calculation steps in CalculateDepositionVelocity()
  are:

  1. Calculate meteorological basics (e.g. viscosity and density of air) and
  whether its day of night (based on parameter SunAngle_degree)

  2. Calculate the Pasquill class (general classification of atmospheric
  stability) based on data at the anemometer location (e.g. grassland land
  use class). This classification is also applied to the target land use
  type.

  3. Calculate other meteorological parameters for the anemometer site
  (Obukhov length, friction velocity, stability corrections) and extrapolate
  wind speed to the blending height, using CalculateWindSpeedAtTargetHeight()

  4. Calculate meteorological parameters for the target land use type (Obukhov
  length, friction velocity, stability corrections), based on the wind speed
  at blending height and the surface properties of the target land use type.

  5. Calculate the aerodynamic resistance (Ra) between the reference height
  and the effective height of the receptor / target land use type. Calculate
  surface resistance (Rs) and gravitation settling.

  6. Calculate dry deposition velocity to the target land use class.
}
\examples{
# Define some standard conditions
Basics <- data.frame(
  SunAngle_degree = 30,
  T_air_K = 293.15,
  AirPressure_Pa = 101325,
  SurfaceIsWet_bool = F,
  GlobalRadiation_W_m2 = 100,
  RelHum_percent = 80,
  CloudCover_percent = 80,
  RoughnessLengthAnemometer_m = 0.03, # WMO grassland
  ZeroPlaneDisplacementHeightAnemometer_m = 0.21,
  AnemometerHeight_m = 10,
  WindSpeedAtAnemometerHeight_ms = 5,
  WindSpeedBlendingHeight_m = 50, # WS no longer dependent on LUC at 50m
  Season = 1,
  DryParticleDiameter_m = 5e-6,
  ParticleDensity_kgm3 = 2000,
  AerosolType = "Dry", # disable hygroscopic swelling
  Parametrization = "Emerson20",
  ReferenceHeight_m = 50 # assuming concentrations are known at blending height
)

# Define two land use classes for which to calculate dry deposition velocity
TargetLUCs <- data.frame(
  # 1=evegreen needleleaf forest, 2=evergreen broadleaf forest
  TargetLUCCodeZhang2001 = c(1, 2),
  RoughnessLengthTargetLUC_m = c(0.8, 2.65) # Zhang01 table 3
) \%>\%
  mutate(
    ZeroPlaneDisplacementHeightTargetLUC_m = 7 * RoughnessLengthTargetLUC_m
  )

# Run calculations
InputTable <- merge(Basics, TargetLUCs)
Output <- CalculateDepositionVelocity2(
  InputTable = InputTable
)

# Show results. LUC 2 (evergreen broadleaf forest) has a higher vd compared to
# LUC 1 (evegreen needleleaf forest) for the specific settings used.
print(Output \%>\% select(TargetLUCCodeZhang2001, V_d_RefHeight_ms))

}
\references{
Zhang L, Gong S, Padro J, Barrie L. A size-segregated particle
dry deposition scheme for an atmospheric aerosol module. Atmospheric
Environment 2001;35:549–560. Emerson EW, Hodshire AL, DeBolt HM, Bilsback KR,
Pierce JR, McMeeking GR, Farmer DK. Revisiting particle dry deposition and
its role in radiative effect estimates. Proceedings of the National Academy
of Sciences 2020;117:26076–26082.
}
