% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateDynamicViscosityOfAir.R
\name{CalculateDynamicViscosityOfAir}
\alias{CalculateDynamicViscosityOfAir}
\title{CalculateDynamicViscosityOfAir}
\usage{
CalculateDynamicViscosityOfAir(T_air_Kelvin)
}
\arguments{
\item{T_air_K}{Air temperature in Kelvin.}
}
\value{
Dynamic viscosity of air in kg/(m*s).
}
\description{
Calculates Dynamic viscosity of air according to Seinfeld and
Pandis (2006) page 909.
}
\references{
Seinfeld JH, Pandis SN. Atmospheric Chemistry and Physics: From Air Pollution to Climate Change. Wiley; 2006.
}
