% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateFrictionVelocity.R
\name{CalculateFrictionVelocity}
\alias{CalculateFrictionVelocity}
\title{CalculateFrictionVelocity}
\usage{
CalculateFrictionVelocity(
  WindSpeed_ms,
  AnemometerHeight_m,
  ZeroPlaneDisplacementHeight_m,
  RoughnessLength_m,
  MoninObukhovLength_m
)
}
\arguments{
\item{WindSpeed_ms}{Wind speed in m/s.}

\item{ZeroPlaneDisplacementHeight_m}{Displacement height in m.}

\item{RoughnessLength_m}{Roughness length in m.}

\item{MoninObukhovLength_m}{Monin-Obkukhiv length in m. Monin-Obkukhiv length for neutral stratification (Pasquill class D) is "infinity". This case is encoded by a value defined in GetConstants()$InfLength in this package.}
}
\value{
A friction velocity value in m.
}
\description{
Calculates friction velocity according to Erisman and Draaijers (1995) page 67 equation 3.25.
}
\examples{

# Friction velocity over grassland for extremely unstable stratification
# (Pasquill class A)
MOL_m <- CalculateMoninObukhovLength(
  PasquillClass = "A",
  RoughnessLength_m = 0.03
)
CalculateFrictionVelocity(
  WindSpeed_ms = 1.5,
  AnemometerHeight_m = 10,
  ZeroPlaneDisplacementHeight_m = 0.21,
  RoughnessLength_m = 0.03,
  MoninObukhovLength_m = MOL_m
)

# Friction velocity over grassland for neutral stratification
# (Pasquill class D)
MOL_m <- CalculateMoninObukhovLength(
  PasquillClass = "D",
  RoughnessLength_m = 0.03
)
CalculateFrictionVelocity(
  WindSpeed_ms = 7,
  AnemometerHeight_m = 10,
  ZeroPlaneDisplacementHeight_m = 0.21,
  RoughnessLength_m = 0.03,
  MoninObukhovLength_m = MOL_m
)
}
\references{
Erisman JW, Draaijers GPJ. Atmospheric Deposition In Relation to Acidification and Eutrophication. 1995.
}
