% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateHygroscopicSwelling.R
\name{CalculateHygroscopicSwelling}
\alias{CalculateHygroscopicSwelling}
\title{CalculateHygroscopicSwelling}
\usage{
CalculateHygroscopicSwelling(
  DryParticleDiameter_m,
  AerosolType,
  RelHum_percent,
  UseCorrectedFormula
)
}
\arguments{
\item{DryParticleDiameter_m}{The diameter of the particles before application
of hygroscopic swelling (dry) in m.}

\item{AerosolType}{Hygroscopic swelling differs depending on aerosol type.
Implemented types are (1) "Dry" for no swelling, (2) "SeaSalt", (3)  "Urban",
(4) "Rural" and (5) "AmmoniumSulfate".}

\item{RelHum_percent}{Relative humidity in percent.}

\item{UseCorrectedFormula}{A boolean indicating whether to use the
(potentially wrong) formulation of hygroscopic swelling according to eq. 10
in Zhang et al. (2001) (UseCorrectedFormula = F) or whether to use the
original formulation according to Gerber 1985 eq. 15 and 22
(UseCorrectedFormula = T). See Details.}
}
\value{
A vector of particle diameters after hygroscopic swelling in m.
}
\description{
Calculates increase in particle diameter due to water update
according to Zhang et al. (2001) eq. 10 or Gerber (1985), see Details.
All parameters must be vectors of same lengths.
}
\details{
Zhang et al. (2001) provides Gerber (1985) as reference for the
formulation of hygroscopic swelling (eq. 10). However, the formulations in
Zhang et al. (2001) eq. 10 and Gerber (1985) (eq. 15 and 22) are not equal.
Gerber (1985) raises the whole equation used by Zhang et al. (2001) to the
power of 1/3. Therefore, if "UseCorrectedFormula = TRUE", the whole
expression from Zhang et al. (2001) is raised to the power of 1/3.
}
\references{
- Zhang L, Gong S, Padro J, Barrie L. A size-segregated particle dry deposition
scheme for an atmospheric aerosol module. Atmospheric Environment
2001;35:549–560.

- Gerber HE. 1985. Relative - Humidity Parameterization of the Navy Aerosol
Model (NAM). NAVAL RESEARCH LAB WASHINGTON DC; December 30, 1985. Available
at: https://apps.dtic.mil/sti/citations/ADA163209.
}
