% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateKinematicViscosityOfAir.R
\name{CalculateKinematicViscosityOfAir}
\alias{CalculateKinematicViscosityOfAir}
\title{CalculateKinematicViscosityOfAir}
\usage{
CalculateKinematicViscosityOfAir(DynamicViscosityAir_kgms, AirDensity_kgm3)
}
\arguments{
\item{DynamicViscosityAir_kgms}{Dynamic viscosity of air in kg/(m*s).}

\item{AirDensity_kgm3}{Density of air in kg/m3.}
}
\value{
Kinematic viscosity of air in kg/(m*s).
}
\description{
Calculates kinematic viscosity of air according to Dixon (2007)
  appendix B
}
\references{
Dixon JC. The Shock Absorber Handbook. John Wiley & Sons; October
  22, 2007.
}
