% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateLossEfficiencyBrownianDiffusion.R
\name{CalculateLossEfficiencyBrownianDiffusion}
\alias{CalculateLossEfficiencyBrownianDiffusion}
\title{CalculateLossEfficiencyBrownianDiffusion}
\usage{
CalculateLossEfficiencyBrownianDiffusion(
  SchmidtNumber,
  BrownianDiffusionParameterGamma,
  Parametrization
)
}
\arguments{
\item{SchmidtNumber}{Schmidt number. E.g. provided by
CalculateSchmidtNumber()}

\item{BrownianDiffusionParameterGamma}{Empirical parameter. Land use specific
in Zhang et al. (2001) and constant at a value of 2/3 in Emerson et al.
(2020)}

\item{Parametrization}{A character defining which parametrization to use.
See ?GetLandUseParameters for a list of valid values.}
}
\value{
Loss efficiency by brownian diffusion
}
\description{
Calculates loss efficiency by brownian diffusion according
to Emerson et al. (2020) eq. 3.
}
\references{
Emerson EW, Hodshire AL, DeBolt HM, Bilsback KR, Pierce JR,
McMeeking GR, Farmer DK. Revisiting particle dry deposition and its role in
radiative effect estimates. Proceedings of the National Academy of Sciences
2020;117:26076–26082.
}
