% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateLossEfficiencyImpaction.R
\name{CalculateLossEfficiencyImpaction}
\alias{CalculateLossEfficiencyImpaction}
\title{CalculateLossEfficiencyImpaction}
\usage{
CalculateLossEfficiencyImpaction(
  StokesNumber,
  ImpactionParameterAlpha,
  Parametrization
)
}
\arguments{
\item{StokesNumber}{Stokes number}

\item{ImpactionParameterAlpha}{A land-use specific empirical paramete}

\item{Parametrization}{A character defining which parametrization to use.
Valid values are "Emerson20" and "Zhang01"}
}
\value{
Stokes number
}
\description{
Calculates loss efficiency by impaction according
to Emerson et al. (2020) eq. 4.
}
\references{
Zhang L, Gong S, Padro J, Barrie L. A size-segregated particle
dry deposition scheme for an atmospheric aerosol module.
Atmospheric Environment 2001;35:549–560.
}
