% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateLossEfficiencyInterception.R
\name{CalculateLossEfficiencyInterception}
\alias{CalculateLossEfficiencyInterception}
\title{CalculateLossEfficiencyInterception}
\usage{
CalculateLossEfficiencyInterception(
  ParticleDiameter_m,
  CharacteristicRadius_m,
  Parametrization,
  SurfaceIsVegetated
)
}
\arguments{
\item{ParticleDiameter_m}{Particle diameter in m.}

\item{CharacteristicRadius_m}{Characteristic radius of the receptor surfaces
in m.}

\item{Parametrization}{A character defining which parametrization to use.
See ?GetLandUseParameters for a list of valid values.}

\item{SurfaceIsVegetated}{A boolean indicated whether the surface is
vegetated. Interception is zero for non-vegetated surfaces.}
}
\value{
Loss efficiency by interception.
}
\description{
Calculates the loss efficiency by interception (E_In) according
to Emerson et al. (2020)
}
\references{
Emerson EW, Hodshire AL, DeBolt HM, Bilsback KR, Pierce JR,
 McMeeking GR, Farmer DK. Revisiting particle dry deposition and its role in
 radiative effect estimates. Proceedings of the National Academy of Sciences
 2020;117:26076–26082.
}
