% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateMeanFreePath.R
\name{CalculateMeanFreePath}
\alias{CalculateMeanFreePath}
\title{CalculateMeanFreePath}
\usage{
CalculateMeanFreePath(T_air_K, AirPressure_Pa, DynamicViscosityAir_kgms)
}
\arguments{
\item{T_air_K}{Air temperature in Kelvin.}

\item{AirPressure_Pa}{Air pressure in Pa.}

\item{DynamicViscosityAir_kgms}{Dynamic viscosity of air in kg/(m*s). E.g. provided by CalculateDynamicViscosityOfAir().}
}
\value{
Mean free path of an air molecule in m.
}
\description{
Calculates the mean free path of an air molecule according to
Seinfeld and Pandis (2006) page 399 eq. 9.6.
}
\examples{
#Validation against example Seinfeld and Pandis eq 9.7 page 399
T_air_K <- 298
AirPressure_Pa = 101325
DynamicViscosityAir_kgms = 1.8e-5
MeanFreePath_m <- CalculateMeanFreePath(T_air_K,AirPressure_Pa,DynamicViscosityAir_kgms)
MeanFreePath_um <- MeanFreePath_m * 1e6
print(MeanFreePath_um)
#0.0651 um as given in SP06

}
\references{
Zhang L, Gong S, Padro J, Barrie L. A size-segregated particle dry deposition scheme for an atmospheric aerosol module. Atmospheric Environment 2001;35:549–560.
}
