% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateMoninObukhovLength.R
\name{CalculateMoninObukhovLength}
\alias{CalculateMoninObukhovLength}
\title{CalculateMoninObukhovLength}
\usage{
CalculateMoninObukhovLength(PasquillClass, RoughnessLength_m)
}
\arguments{
\item{PasquillClass}{A single character indicating the Pasquill stability class. For example according to Seinfeld and Pandis
(2006) page 750 as implemented in GetPasquillClass().}

\item{RoughnessLength_m}{Roughness length in m.}
}
\value{
A numeric value for Monin-Obukhov length. Monin-Obkukhiv length for neutral stratification (Pasquill class D) is "infinity". This case is encoded by a value defined in GetConstants()$InfLength in this package.
}
\description{
Calculates Monin-Obukhov length according to Seinfeld and Pandis
(2006) page 751 eq. 16.83.
}
\examples{

# For Pasquill class A ("extremely unstable" conditions) and grassland
CalculateMoninObukhovLength(
  PasquillClass = "A",
  RoughnessLength_m = 0.03
)

# For Pasquill class D ("neutral" conditions) and grassland
CalculateMoninObukhovLength(
  PasquillClass = "D",
  RoughnessLength_m = 0.03
)
GetConstants()$InfLength

}
\references{
Seinfeld JH, Pandis SN. Atmospheric Chemistry and Physics: From Air Pollution to Climate Change 2006.
}
