% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSettlingVelocity.R
\name{CalculateSettlingVelocity}
\alias{CalculateSettlingVelocity}
\title{CalculateSettlingVelocity}
\usage{
CalculateSettlingVelocity(
  ParticleDensity_kgm3,
  ParticleDiameter_m,
  MeanFreePathOfAirMolecule_m,
  DynamicViscosityAir_kgms
)
}
\arguments{
\item{ParticleDensity_kgm3}{Particle density in kg/m3.}

\item{ParticleDiameter_m}{Particle diameter in m.}

\item{MeanFreePathOfAirMolecule_m}{Mean free path of an air molecule in m. E.g. provided by CalculateMeanFreePath().}

\item{DynamicViscosityAir_kgms}{Dynamic viscosity of air in kg/(m*s). E.g. provided by CalculateDynamicViscosityOfAir().}
}
\value{
Gravitation settling velocity in m/s.
}
\description{
Calculates settling velocity of particles by gravitation according to Zhang et al. (2001).
}
\references{
Zhang L, Gong S, Padro J, Barrie L. A size-segregated particle dry deposition scheme for an atmospheric aerosol module. Atmospheric Environment 2001;35:549–560.
}
