% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSurfaceResistance.R
\name{CalculateSurfaceResistance}
\alias{CalculateSurfaceResistance}
\title{CalculateSurfaceResistance}
\usage{
CalculateSurfaceResistance(
  SurfaceIsWet,
  FrictionVelocity_ms,
  StokesNumber,
  E_b,
  E_Im,
  E_In,
  Parametrization
)
}
\arguments{
\item{SurfaceIsWet}{Indicator whether the receptor surface is wet (for
bounce correction term), boolean.}

\item{FrictionVelocity_ms}{Friction velocity in m/s.}

\item{StokesNumber}{Stokes number.
in m.}

\item{E_b}{Loss efficiency by brownian diffusion.
in m.
#'}

\item{E_Im}{Loss efficiency by impaction.
in m.
#'}

\item{E_In}{Loss efficiency by interception.
in m.}

\item{Parametrization}{A character defining which parametrization to use.
Valid values are "Emerson20" and "Zhang01"}
}
\value{
Surface resistance in s/m.
}
\description{
Calculates the surface resistance (R_s) according to Zhang et
al. (2001).
}
\references{
Zhang L, Gong S, Padro J, Barrie L. A size-segregated particle
dry deposition scheme for an atmospheric aerosol module. Atmospheric
Environment 2001;35:549–560.
}
