% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateWindSpeedAtTargetHeight.R
\name{CalculateWindSpeedAtTargetHeight}
\alias{CalculateWindSpeedAtTargetHeight}
\title{CalculateWindSpeedAtTargetHeight}
\usage{
CalculateWindSpeedAtTargetHeight(
  FrictionVelocity_ms,
  TargetHeight_m,
  ZeroPlaneDisplacementHeight_m,
  RoughnessLength_m,
  MoninObukhovLength_m
)
}
\arguments{
\item{FrictionVelocity_ms}{Friction velocity in m/s.}

\item{TargetHeight_m}{Height at which the wind speed should be calculated in
m.}

\item{ZeroPlaneDisplacementHeight_m}{Zero-plane displacement height in m.}

\item{RoughnessLength_m}{Roughness length in m.}

\item{MoninObukhovLength_m}{Monin-Obukhov-length in m.}
}
\value{
Wind speed at target height in m/s.
}
\description{
Calculate wind speed according to stability-corrected log wind
profile. This function is rearrangement of the formula for friction velocity
after Erisman and Draaijers (1995) page 67. It is in line with
https://en.wikipedia.org/wiki/Log_wind_profile
}
\references{
Erisman JW, Draaijers GPJ. Atmospheric Deposition In Relation to Acidification
}
