% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dd_subprocess_validation.R
\docType{data}
\name{dd_subprocess_validation}
\alias{dd_subprocess_validation}
\title{Data for validation of dry deposition subprocesses}
\format{
A data frame with 153 rows and 4 columns:
\describe{
  \item{ParticleDiameter_um}{Particle diameter in um}
  \item{DepositionVelocity_cms}{Dry deposition velocity resulting from the
  respective dry-deposition sub-process in cm^2/s}
  \item{Process}{Dry-deposition sub-process}
  \item{Parametrization}{Parametrization (GEOS-Chem old or GEOS-Chem new)}
}
}
\usage{
data(dd_subprocess_validation)
}
\description{
Data extracted from Emerson et al. (2020) fig. 2. Shows the dependency of
dry deposition subprocesses to particle size according to the old and the old
and the revised parametrization of the GEOS-Chem model.
Note that the old parametrization is not identical to the Zhang et al. (2001)
parametrization although it is termed "Zhang (2001)" in the Emerson et al.
(2020) paper. See ?GetLandUseParameters for details.
}
\references{
Emerson EW, Hodshire AL, DeBolt HM, Bilsback KR, Pierce JR,
McMeeking GR, Farmer DK. Revisiting particle dry deposition and its role in
radiative effect estimates. Proceedings of the National Academy of Sciences
2020;117:26076–26082.
}
\keyword{datasets}
