% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dd_validation.R
\docType{data}
\name{dd_validation}
\alias{dd_validation}
\title{Data for validation of dry deposition}
\format{
A data frame with 106 rows and 4 columns: \describe{
  \item{ParticleDiameter_um}{Particle diameter in um}
  \item{DepositionVelocity_cms}{Dry deposition velocity in cm^2/s}
  \item{Parametrization}{Parametrization (GEOS-Chem old (GCOld) or GEOS-Chem new (GCNew))}
  \item{LUCName}{Land use class name}
  }
}
\usage{
data(dd_validation)
}
\description{
Data extracted from Emerson et al. (2020) fig. 1. Shows the dependency of
dry deposition velocity to particle size for different land use types
according to the old and the revised parametrization of the GEOS-Chem model.
Note that the old parametrization is not identical to the Zhang et al. (2001)
parametrization although it is termed "Zhang (2001)" in the Emerson et al.
(2020) paper. See ?GetLandUseParameters for details.
}
\references{
Emerson EW, Hodshire AL, DeBolt HM, Bilsback KR, Pierce JR,
  McMeeking GR, Farmer DK. Revisiting particle dry deposition and its role in
  radiative effect estimates. Proceedings of the National Academy of Sciences
  2020;117:26076–26082.

  Zhang L, Gong S, Padro J, Barrie L. A size-segregated particle dry
  deposition scheme for an atmospheric aerosol module. Atmospheric
  Environment 2001;35:549–560.
}
\keyword{datasets}
