% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_diffusion_validation.R
\docType{data}
\name{diffusion_validation}
\alias{diffusion_validation}
\title{Data for validation of the diffusion term}
\format{
A data frame with 38 rows and 2 columns:
\describe{
  \item{ParticleDiameter_um}{Particle diameter in um}
  \item{DiffusionCoefficient_cm2s}{Diffusion coefficient in cm^2/s}
}
}
\usage{
data(diffusion_validation)
}
\description{
Shows the dependency of the diffusion coefficient and particle diameter.
Data extracted from Seinfeld and Pandis (2006) page 417 fig. 9.8.
}
\references{
Seinfeld JH, Pandis SN. Atmospheric Chemistry and Physics: From
Air Pollution to Climate Change. Wiley; 2006.
}
\keyword{datasets}
