% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLandUseParametersZhang2001.R
\name{GetLandUseParametersZhang2001}
\alias{GetLandUseParametersZhang2001}
\title{GetLandUseParametersZhang2001}
\usage{
GetLandUseParametersZhang2001(LUCs, Seasons, TargetPar)
}
\arguments{
\item{LUCs}{A vector of land use class codes (integer values). Currently, only land use classes 1-7 are implemented.}

\item{Seasons}{A vector of season codes (integer values 1-5).}

\item{TargetPar}{A string indicating which parameter to return from Zhang et al. (2001) table 3 ("z_0", "A", "alpha" or "gamma").}
}
\value{
A vector of values for parameter "TargetPar".
}
\description{
Get land use parameters according to Zhang et al. (2001) table 3. The function is vectorized with respect to parameters "LUC" and "Seasons". I.e. these two parameters must be vectors of same length.
}
\examples{
GetLandUseParametersZhang2001(LUCs = c(1,2), Seasons = c(2,5), TargetPar = "A")
}
\references{
Zhang L, Gong S, Padro J, Barrie L. A size-segregated particle dry deposition scheme for an atmospheric aerosol module. Atmospheric Environment 2001;35:549–560.
}
