% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetParameters.R
\name{GetParameters}
\alias{GetParameters}
\title{GetParameters}
\usage{
GetParameters(Parametrization, TargetParameter)
}
\arguments{
\item{Parametrization}{A character defining which parametrization to use.
See ?GetLandUseParameters for a list of valid values.}

\item{TargetParameter}{A character indicating which parameter value to
return. Valid options are "C_b", "beta", "C_Im", "nu", "C_In" and
"epsilon_0". See Emerson et al. (2020) table S1.}
}
\value{
A named list of parameters.
}
\description{
This function returns empirical constants for dry deposition
  sub-processes according to Zhang et al. (2001) or according to the
  re-paramtrization by Emerson et al. (2020). This function covers only
  parameters that are not land-use specific (see GetLandUseParameters()).
}
\examples{
GetParameters(Parametrization = "Zhang01", TargetParameter = "C_b")
GetParameters(Parametrization = "GCNewSeason", TargetParameter = "C_b")

}
\references{
Emerson EW, Hodshire AL, DeBolt HM, Bilsback KR, Pierce JR,
McMeeking GR, Farmer DK. Revisiting particle dry deposition and its role in
radiative effect estimates. Proceedings of the National Academy of Sciences
2020;117:26076–26082.

Zhang L, Gong S, Padro J, Barrie L. A size-segregated particle dry deposition
scheme for an atmospheric aerosol module. Atmospheric Environment
2001;35:549–560.
}
