% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPasquillClass.R
\name{GetPasquillClass}
\alias{GetPasquillClass}
\title{GetPasquillClass}
\usage{
GetPasquillClass(
  SurfaceWindSpeed_ms,
  DayOrNight,
  IncomingSolarRadiation_Wm2,
  CloudCover_percent
)
}
\arguments{
\item{SurfaceWindSpeed_ms}{Wind speed at surface in m/s.}

\item{DayOrNight}{Boolean indicating whther it is day or night.}

\item{IncomingSolarRadiation_Wm2}{Solar radiation in W/m2.}

\item{CloudCover_percent}{Cloud cover in percent.}
}
\value{
Chracter string indicating the Pasquill stability class (A - F)
}
\description{
Calculates Pasquill stability class according to Seinfeld and
Pandis (2006) page 750.
}
\references{
Seinfeld JH, Pandis SN. Atmospheric Chemistry and Physics: From
Air Pollution to Climate Change. 2006.
}
