% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_land_use_parameters_zhang01.R
\docType{data}
\name{land_use_parameters_zhang01}
\alias{land_use_parameters_zhang01}
\title{Land-use specific parameters according to Zhang et al. (2001) table 3.}
\format{
A data frame with 120 rows and 4 columns:
\describe{
  \item{Parameter}
  \item{Season}
  \item{LUC} {Land use class}
  \item{value}
}
}
\usage{
data(land_use_parameters_zhang01)
}
\description{
Note that parameter A (characteristic radius of receptor surface) comes in mm
and not in m. Note that the Emerson et al. (2020) parametrization uses a
different value for parameter gamma. Season 999 indicates that the value
applies to all seasons.
}
\references{
Zhang L, Gong S, Padro J, Barrie L. A size-segregated particle
dry deposition scheme for an atmospheric aerosol module. Atmospheric
Environment 2001;35:549–560.

Emerson EW, Hodshire AL, DeBolt HM, Bilsback KR, Pierce JR, McMeeking GR,
Farmer DK. Revisiting particle dry deposition and its role in radiative
effect estimates. Proceedings of the National Academy of Sciences
2020;117:26076–26082.
}
\keyword{datasets}
