% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_meteo_time_series.R
\docType{data}
\name{meteo_time_series}
\alias{meteo_time_series}
\title{Example 48 h meteorological time series}
\format{
A data frame with 48 rows and 11 columns
}
\usage{
data(meteo_time_series)
}
\description{
Meteorological data required for dry deposition modelling (e.g. to determine
atmospheric stability via Pasquill classes). Hourly data is for example
available from the ERA5 model
<https://cds.climate.copernicus.eu/cdsapp#!/dataset/reanalysis-era5-land?tab=overview>
}
\keyword{datasets}
