% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_land_use_parameters_zhang01.R,
%   R/data_vd_windspeed_validation_zhang01.R
\docType{data}
\name{vd_windspeed_validation_zhang01}
\alias{vd_windspeed_validation_zhang01}
\title{Land-use specific parameters according to Zhang et al. (2001) table 3.}
\format{
A data frame with 120 rows and 4 columns:
\describe{
  \item{Parameter}
  \item{Season}
  \item{LUC} {Land use class}
  \item{value}
}

A data frame with 126 rows and 3 columns:
\describe{
  \item{ParticleDiameter_um}{Particle diameter in um}
  \item{DepositionVelocity_cms}{Dry deposition velocity in cm/s}
  \item{WindSpeed_ms}{Wind speed in m/s}
}
}
\usage{
data(land_use_parameters_zhang01)

data(vd_windspeed_validation_zhang01)
}
\description{
Note that parameter A (characteristic radius of receptor surface) comes in mm
and not in m. Note that the Emerson et al. (2020) parametrization uses a
different value for parameter gamma. Season 999 indicates that the value
applies to all seasons.

Shows the dependency of vd to particle size for different wind speeds for
the Zhang et al. (2001) parametrization. Data extracted from Zhang et al.
fig. 2 panel "Needleleaf trees". According to Zhang et al. (2001) and Gong et
al. (2003), the data applies for the following settings: Neutral stability
(PasquillClass = "D"), AnemometerHeight = 20 m, ReferenceHeight = 10 m.
}
\references{
Zhang L, Gong S, Padro J, Barrie L. A size-segregated particle
dry deposition scheme for an atmospheric aerosol module. Atmospheric
Environment 2001;35:549–560.

Emerson EW, Hodshire AL, DeBolt HM, Bilsback KR, Pierce JR, McMeeking GR,
Farmer DK. Revisiting particle dry deposition and its role in radiative
effect estimates. Proceedings of the National Academy of Sciences
2020;117:26076–26082.

Zhang L, Gong S, Padro J, Barrie L. A size-segregated particle
dry deposition scheme for an atmospheric aerosol module. Atmospheric
Environment 2001;35:549–560.

Gong SL, Barrie LA, Blanchet J-P, von Salzen K, Lohmann U, Lesins G, Spacek
L, et al. Canadian Aerosol Module: A size-segregated simulation of
atmospheric aerosol processes for climate and air quality models 1. Module
development. Journal of Geophysical Research:
Atmospheres 2003;108:AAC 3-1-AAC 3-16.
}
\keyword{datasets}
