% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateDepositionVelocity.R
\name{CalculateDepositionVelocity}
\alias{CalculateDepositionVelocity}
\title{CalculateDepositionVelocity}
\usage{
CalculateDepositionVelocity(InputTable)
}
\arguments{
\item{InputTable}{A data frame with the following columns:

 SunAngle_degree: The sun angle in degree. Only used to determine if its "day" (> 0 degree) or "night" (< 0 degree) which is required for the determination of the Pasquill stability class (ddpart::GetPasquillClass()). The sun angle can easily be calculated with oce::sunAngle().

 T_air_K: Air temperature in Kelvin.

 AirPressure_Pa: Air pressure in Pa.

 GlobalRadiation_W_m2: Global radiation in W/m2.

 CloudCover_percent

 RelHum_percent:

 WindSpeedAtAnemometerHeight_ms:

 RoughnessLengthAnemometer_m:

 ZeroPlaneDisplacementHeightAnemometer_m:

 AnemometerHeight_m:

 WindSpeedBlendingHeight_m:

 Season:

 TargetLUCCodeZhang2001:

 ReferenceHeight_m:

 RoughnessLengthTargetLUC_m:

 ZeroPlaneDisplacementHeightTargetLUC_m:

 DryParticleDiameter_m:

 ParticleDensity_kgm3:

 AerosolType: Set parameter AerosolType to "Dry" to disable calculation of hygroscopic swelling. Set parameter "AerosolType" to one of "SeaSalt", "Urban", "Rural", "AmmoniumSulfate" to enable hygroscopic swelling (see CalculateWetRadius()).}
}
\value{
A data frame repeating the InputTable plus additional columns with calculated values.
}
\description{
Calculates the deposition velocity of particle according to the Zhang et al. (2001) and the Emerson et al. 2020 publications. Required inputs are:

(1) basic meteorological data (wind speed, global ration, relative humidity, cloud cover, air temperature, air pressure, surface wetness)

(2) information on the surface properties (land use class, roughness length and displacement height) and reference height (height of concentration measurements)

(3) particle properties (dry particle diameter, density, optional: aerosol type to calculate hygroscopic swelling)

The approach is designed for use cases where information on wind speed is not available directly for the target location / target land use type. For example, wind speeds are often modelled for or measured at grassland sites, but the dry deposition velocity is of interest at a close by for forest or cropland site. The approach therefore differentiates between two sites:

- "anemometer site": Location/land use class where the wind speed measurements are available. Characterized by parameters "RoughnessLengthAnemometer_m" and "ZeroPlaneDisplacementHeightAnemometer_m"

- "target land use site": Location/land use class where the dry deposition velocity should be calculated. Characterized by parameters "RoughnessLengthTargetLUC_m", "ZeroPlaneDisplacementHeightTargetLUC_m", "Season"

CalculateDepositionVelocity() vertically extrapolates the wind speed at "anemometer site" the to a "blending height" (e.g. 50 m) where atmospheric conditions are somewhat independent of the underlying surface. From the blending height, the parameters relevant for particle dry deposition to the target land use site (friction velocity, aerodynamic resistance, etc.) are be calculated. The approach can also be used for cases where wind speed data is available for the target land use type directly. The main calculation steps in CalculateDepositionVelocity() are:

1. Calculate meteorological basics (e.g. viscosity and density of air) and whether its day of night (based on parameter SunAngle_degree)

2. Calculate the Pasquill class (general classification of atmospheric stability) based on data at the anemometer location (e.g. grassland land use class). This classification is also applied to the target land use type.

3. Calculate other meteorological parameters for the anemometer site (Obukhov length, friction velocity, stability corrections) and extrapolate wind speed to the blending height.

- Calculate meteorological parameters for the target land use type (Obukhov length, friction velocity, stability corrections), based on the wind speed at blending height and the surface properties of the target land use type.

- Calculate aerodynamic resistance (Ra) between the reference height and the receptor and surface resistance (Rs)
}
\examples{
See vignette.
}
\references{
Zhang L, Gong S, Padro J, Barrie L. A size-segregated particle dry deposition scheme for an atmospheric aerosol module. Atmospheric Environment 2001;35:549–560.
Emerson EW, Hodshire AL, DeBolt HM, Bilsback KR, Pierce JR, McMeeking GR, Farmer DK. Revisiting particle dry deposition and its role in radiative effect estimates. Proceedings of the National Academy of Sciences 2020;117:26076–26082.
}
