% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSchmidtNumber.R
\name{CalculateSchmidtNumber}
\alias{CalculateSchmidtNumber}
\title{CalculateSchmidtNumber}
\usage{
CalculateSchmidtNumber(
  DynamicViscosityAir_kgms,
  KinematicViscosityOfAir_m2s,
  T_air_K,
  ParticleDiameter_m
)
}
\arguments{
\item{DynamicViscosityAir_kgms}{Dynamic viscosity of air in kg/(m*s). E.g.
provided by CalculateDynamicViscosityOfAir().}

\item{KinematicViscosityOfAir_m2s}{Kinematic viscosity of air in m2/2. E.g.
provided by CalculateKinematicViscosityOfAir().}

\item{T_air_K}{Air temperature in Kelvin.}

\item{ParticleDiameter_m}{Particle diameter in m.}
}
\value{
Schmidt number
}
\description{
Calculates the Schmidt number according to Seinfeld and Pandis
  (2006) page 574
}
\examples{

# Reproduce relation between brownian diffusivity and particle diameter as shown
# in Seinfeld and Pandis (2006) page 417 Figure 9.8

data(diffusion_validation)
PlotData <- data.frame(
  #Set standard atmospheric conditions
  T_air_K = 20,
  AirPressure_Pa = 101325,
  #Set particle diameter according to Seinfeld and Pandis (2006)
  ParticleDiameter_m = 10^seq(-9, -5, length.out = 50)
) \%>\%
  #Calculation of quantities related to diffusion
  mutate(
    DynamicViscosityAir_kgms = CalculateDynamicViscosityOfAir(T_air_K),
    AirDensity_kgm3 = CalculateAirDensity(
      AirPressure_Pa = AirPressure_Pa,
      T_air_K = T_air_K
    ),
    KinematicViscosityOfAir_m2s = CalculateKinematicViscosityOfAir(
      DynamicViscosityAir_kgms = DynamicViscosityAir_kgms,
      AirDensity_kgm3 = AirDensity_kgm3
    ),
    SchmidtNumber = CalculateSchmidtNumber(
      DynamicViscosityAir_kgms = DynamicViscosityAir_kgms,
      KinematicViscosityOfAir_m2s = KinematicViscosityOfAir_m2s,
      T_air_K = T_air_K,
      ParticleDiameter_m = ParticleDiameter_m
    ),
    # Calculate brownian diffusivity from intermediate results
    BrownianDiffusivity_m2s = KinematicViscosityOfAir_m2s / SchmidtNumber,
    # For plotting:
    BrownianDiffusivity_cm2s = BrownianDiffusivity_m2s * 1e4,
    ParticleDiameter_um = ParticleDiameter_m * 1e6,
    Type = "Results from ddpart"
  )

#Plot
if (require("ggplot2")) {
ggplot()  +
  geom_line(
    data = diffusion_validation \%>\%
      mutate(
        Type = "Validation data from Seinfeld and Pandis (2006)"
      ),
    mapping = aes(
      x = log10(ParticleDiameter_um),
      y = log10(DiffusionCoefficient_cm2s),
      color = Type
    )
  ) +
  geom_point(
    data = PlotData,
    mapping = aes(
      x = log10(ParticleDiameter_um),
      y = log10(BrownianDiffusivity_cm2s),
      color = Type
    ),
    shape = "x"
  ) +
  annotate(
    geom = "text",
    x = -0,
    y = -3,
    label = "Small differences likely related\nto errors from extraction of
    validation data from PDF figure."
  ) +
  theme(
    legend.position = "bottom"
  )
 }

}
\references{
Seinfeld JH, Pandis SN. Atmospheric Chemistry and Physics: From
  Air Pollution to Climate Change. Wiley; 2006.
}
