% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSchmidtNumber.R
\name{CalculateSchmidtNumber}
\alias{CalculateSchmidtNumber}
\title{CalculateSchmidtNumber}
\usage{
CalculateSchmidtNumber(
  DynamicViscosityAir,
  KinematicViscosityOfAir_m2s,
  T_air_K,
  ParticleDiameter_m
)
}
\arguments{
\item{KinematicViscosityOfAir_m2s}{Kinematic viscosity of air in m2/2. E.g. provided by CalculateKinematicViscosityOfAir().}

\item{T_air_K}{Air temperature in Kelvin.}

\item{ParticleDiameter_m}{Particle diameter in m.}

\item{DynamicViscosityAir_kgms}{Dynamic viscosity of air in kg/(m*s). E.g. provided by CalculateDynamicViscosityOfAir().}
}
\value{
Schmidt number
}
\description{
Calculates the Schmidt number according to Seinfeld and
Pandis (2006) page 574
}
\references{
Seinfeld JH, Pandis SN. Atmospheric Chemistry and Physics: From Air Pollution to Climate Change. Wiley; 2006.
}
