% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSurfaceResistance.R
\name{CalculateSurfaceResistance}
\alias{CalculateSurfaceResistance}
\title{CalculateSurfaceResistance}
\usage{
CalculateSurfaceResistance(
  SurfaceIsWet,
  FrictionVelocity_ms,
  StokesNumber,
  E_b,
  E_Im,
  E_In,
  ParticleDiameter_m,
  Parametrization
)
}
\arguments{
\item{SurfaceIsWet}{Indicator whether the receptor surface is wet (for
bounce correction term), boolean.}

\item{FrictionVelocity_ms}{Friction velocity in m/s.}

\item{StokesNumber}{Stokes number.}

\item{E_b}{Loss efficiency by brownian diffusion.}

\item{E_Im}{Loss efficiency by impaction.}

\item{E_In}{Loss efficiency by interception.}

\item{ParticleDiameter_m}{Particle diameter in m.}

\item{Parametrization}{A character defining which parametrization to use.
See ?GetLandUseParameters for a list of valid values.}
}
\value{
Surface resistance in s/m.
}
\description{
Calculates the surface resistance (R_s) according to Zhang et
al. (2001).
}
\references{
Zhang L, Gong S, Padro J, Barrie L. A size-segregated particle
dry deposition scheme for an atmospheric aerosol module. Atmospheric
Environment 2001;35:549–560.
}
