% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dd_validation.R
\docType{data}
\name{dd_validation}
\alias{dd_validation}
\title{Data for validation of dry deposition}
\format{
A data frame with 106 rows and 4 columns: \describe{
  \item{ParticleDiameter_um}{Particle diameter in um}
  \item{DepositionVelocity_cms}{Dry deposition velocity in cm^2/s}
  \item{Parametrization}{The parametrization for which the data has been
  extracted (Zhang01 or Emerson20)} \item{LUC}{Land use class: 1 (Needleleaf
  forest, 2 broadleaf forest, 6 grassland)} }
}
\usage{
data(dd_validation)
}
\description{
Shows the dependency of dry deposition velocity to particle size for
different land use types according to Emerson et al. (2020) and Zhang et al.
(2001). Data extracted from Emerson et al. (2020) fig. 1.
}
\references{
Emerson EW, Hodshire AL, DeBolt HM, Bilsback KR, Pierce JR,
  McMeeking GR, Farmer DK. Revisiting particle dry deposition and its role in
  radiative effect estimates. Proceedings of the National Academy of Sciences
  2020;117:26076–26082.
}
\keyword{datasets}
