% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetParameters.R
\name{GetParameters}
\alias{GetParameters}
\title{GetParameters}
\usage{
GetParameters(Parametrization, TargetParameter)
}
\value{
A named list of parameters.
}
\description{
This function returns empirical constants for dry deposition
sub-processes according to Zhang et al. (2001) or according to the
re-paramtrization by Emerson et al. (2020). This function covers only
parameters that are not land-use specific (see GetLandUseParameters()).
}
\examples{
GetParameters(Parametrization = "Zhang01")
GetParameters(Parametrization = "Emerson20")

}
\references{
Zhang L, He Z. Technical Note: An empirical algorithm estimating dry deposition velocity of fine, coarse and giant particles. Atmospheric Chemistry and Physics 2014;14:3729–3737.
Zhang L, Gong S, Padro J, Barrie L. A size-segregated particle dry deposition scheme for an atmospheric aerosol module. Atmospheric Environment 2001;35:549–560.
Emerson EW, Hodshire AL, DeBolt HM, Bilsback KR, Pierce JR, McMeeking GR, Farmer DK. Revisiting particle dry deposition and its role in radiative effect estimates. Proceedings of the National Academy of Sciences 2020;117:26076–26082.
}
