% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_vd_windspeed_validation_zhang01.R
\docType{data}
\name{vd_windspeed_validation_zhang01}
\alias{vd_windspeed_validation_zhang01}
\title{Data for validation of vd in relation to particle size and wind speed.}
\format{
A data frame with 126 rows and 3 columns:
\describe{
  \item{ParticleDiameter_um}{Particle diameter in um}
  \item{DepositionVelocity_cms}{Dry deposition velocity in cm/s}
  \item{WindSpeed_ms}{Wind speed in m/s}
}
}
\usage{
data(vd_windspeed_validation_zhang01)
}
\description{
Shows the dependency of vd to particle size for different wind speeds for
the Zhang et al. (2001) parametrization. Data extracted from Zhang et al.
fig. 2 panel "Needleleaf trees". According to Zhang et al. (2001) and Gong et
al. (2003), the data applies for the following settings: Neutral stability
(PasquillClass = "D"), AnemometerHeight = 20 m, ReferenceHeight = 10 m.
}
\references{
Zhang L, Gong S, Padro J, Barrie L. A size-segregated particle
dry deposition scheme for an atmospheric aerosol module. Atmospheric
Environment 2001;35:549–560.

Gong SL, Barrie LA, Blanchet J-P, von Salzen K, Lohmann U, Lesins G, Spacek
L, et al. Canadian Aerosol Module: A size-segregated simulation of
atmospheric aerosol processes for climate and air quality models 1. Module
development. Journal of Geophysical Research:
Atmospheres 2003;108:AAC 3-1-AAC 3-16.
}
\keyword{datasets}
